[IMP] web: graphs in kandan cards
This commit aims to modernize the graphs in the kanban card.

The graphs now fill the entire available space.

For the bar chart, external borders have been removed, and the hover
effect has been change to also highlight the bar if the column is
hovered.

For the line chart, external dots have been removed aswell, and the line
is now curved.

task-5186572

closes odoo/odoo#238675

Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>